only need to fetch peers once before starting listeners
