+++ lib/p2p/connection_handler.ex
@@ -7,15 +7,8 @@ defmodule Elixium.P2P.ConnectionHandler do
   @doc """
     Spawn a new handler, and have it run the authentication code immediately
   """
-  @spec start_link(reference, pid) :: {:ok, pid}
-  def start_link(socket, pid) do
-    # Fetch known peers. We're going to try to connect to them
-    # before setting up a listener
-    peers = [{'localhost', 31013}]
-    # peers = PeerStore.load_known_peers
-
-    peers = find_potential_peers()
-
+  @spec start_link(reference, pid, List) :: {:ok, pid}
+  def start_link(socket, pid, peers) do
     pid =
       case peers do
         :not_found ->
@@ -142,7 +135,7 @@ defmodule Elixium.P2P.ConnectionHandler do
   end
 
   defp generate_session_key(shared_secret) do
-    # Truncate the key to be 32 bytes (256 bits) since AES256 won't accept anything bigger
+    # Truncate the key to be 32 bytes (256 bits) since AES256 won't accept anything bigger.
     # Originally, I was worried this would be a security flaw, but according to
     # https://crypto.stackexchange.com/questions/3288/is-truncating-a-hashed-private-key-with-sha-1-safe-to-use-as-the-symmetric-key-f
     # it isn't
@@ -168,32 +161,5 @@ defmodule Elixium.P2P.ConnectionHandler do
       :not_found -> false
       {_identifier, _password} -> true
     end
-  end
-
-  # Either loads peers from a local storage or connects to the
-  # bootstrapping registry
-  @spec find_potential_peers :: List | :not_found
-  defp find_potential_peers do
-    case PeerStore.load_known_peers do
-      :not_found -> fetch_peers_from_registry
-      peers -> peers
-    end
-  end
-
-  # Connects to the bootstrapping peer registry and returns a list of
-  # previously connected peers.
-  @spec fetch_peers_from_registry :: List
-  defp fetch_peers_from_registry do
-    {:ok, {{'HTTP/1.1', 200, 'OK'}, _headers, body}} = :httpc.request('http://testnet-peer-registry.w3tqcgzmeb.us-west-2.elasticbeanstalk.com/31013')
-    peers =
-      body
-      |> Jason.decode!()
-      |> Enum.map(fn p ->
-        [ip, port] = String.split(p, ":")
-        {port, _} = Integer.parse(port)
-        ip = String.to_charlist(ip)
-
-        {ip, port}
-      end)
   end
 end
